Fix hashtag being off-center in smaller viewports.
The navbar uses flexbox to place its items easily. The problem with this
is that the width of each item is taken into account for the placement
of the next item, and since each item has a different width, the second
item might not always be in the exact center.

The workaround is to set the flex-basis property in smaller viewports to
let flexbox give the same width to each item. This still means that the
menu in the full-width viewport isn't in the exact center, but that
doesn't look too bad.